Move certification-center-invitations controller tests
